UT: Fix and move Agent plugin resources' tests
