+++ monkey/tests/unit_tests/monkey_island/cc/services/agent_plugin_service/flask_resources/test_agent_plugins.py
@@ -11,7 +11,7 @@ from tests.unit_tests.monkey_island.cc.fake_agent_plugin_data import (
 from tests.unit_tests.monkey_island.conftest import get_url_for_resource
 
 from monkey_island.cc.repositories import IAgentPluginRepository, RetrievalError
-from monkey_island.cc.resources import AgentPlugins
+from monkey_island.cc.services.agent_plugin_service.flask_resources import AgentPlugins
 
 FAKE_PLUGIN_NAME = "plugin_abc"
